+++ devito/dse/__init__.py
@@ -1,5 +1,4 @@
 from devito.dse.aliases import *  # noqa
 from devito.dse.clusterizer import *  # noqa
-from devito.dse.graph import *  # noqa
 from devito.dse.manipulation import *  # noqa
 from devito.dse.transformer import *  # noqa
