+++ devito/dse/clusterizer.py
@@ -1,11 +1,10 @@
 from collections import OrderedDict, namedtuple
 
-from devito.dse.graph import temporaries_graph
-
+from devito.ir.dfg import temporaries_graph
 from devito.symbolics import xreplace_indices
-from devito.types import Scalar
 from devito.stencil import Stencil
 from devito.tools import as_tuple
+from devito.types import Scalar
 
 __all__ = ['clusterize', 'optimize']
