+++ devito/ir/__init__.py
@@ -1,0 +1,1 @@
+from devito.ir.dfg import *  # noqa
