+++ devito/ir/dfg/__init__.py
@@ -1,0 +1,1 @@
+from devito.ir.dfg.graph import *  # noqa
